🤖 fix(goal-board): reject partial turn-cap strings + serialize workspace-default saves (Codex P2)
Two follow-ups:

1. **Reject partial turn-cap strings.** The new turn-cap inputs in the
   Upcoming Adder and the inline UpcomingRowEditor used
   `Number.parseInt`, which silently truncates inputs like `1.5` → 1
   and `12abc` → 12. The user would type one value and the queued goal
   would persist a different cap.

   Extracted `parseGoalTurnCapInput` (full-string `Number` +
   `Number.isSafeInteger` validation) to `src/common/utils/goals/
   budgetParser.ts` so every entry point — slash command, GoalTab inline
   editor, Adder, inline editor — validates the same way.
   `GoalTab.tsx`'s local `parseTurnCapInput` is now an alias for the
   shared canonical parser (eliminates the duplicated definition that
   the Codex note flagged as drifting risk).

2. **Serialize workspace-default saves.** Codex P2: rapid edits to two
   workspace defaults fired multiple concurrent full-record `set` calls
   to `workspace.goalDefaults.set`. The backend reads → mutates →
   writes the full config record per request, so a slower first
   request could complete after a later edit and overwrite it on disk
   while the UI showed the optimistic latest state — silent reverts
   on next refetch or restart.

   Fix in `GoalDefaultsSection.tsx`: chain saves through a
   `useRef<Promise<void>>` so each request awaits the previous before
   issuing its own write. The optimistic local `setOverride` still fires
   immediately so the UI feels snappy, but on-disk state is now
   guaranteed to match the user's final edit.

---

_Generated with `mux` • Model: `anthropic:claude-opus-4-7` • Thinking: `xhigh` • Cost: `$179.68`_

<!-- mux-attribution: model=anthropic:claude-opus-4-7 thinking=xhigh costs=179.68 -->
