+++ src/common/utils/goals/budgetParser.ts
@@ -42,3 +42,25 @@ export function parseGoalBudgetInputCents(value: string): number | null | undefi
   const cents = Number((dollarsMatch[2] ?? "").padEnd(2, "0"));
   return dollars * 100 + cents;
 }
+
+/**
+ * Canonical user-facing turn-cap parser. Matches the strict-int semantics
+ * of `GoalTab`'s legacy `parseTurnCapInput` so every entry point (`/goal
+ * budget`, GoalTab inline editor, GoalBoard Adder, GoalBoard inline
+ * editor) accepts the same set of inputs. Partial strings like `1.5` or
+ * `12abc` correctly fail validation here, unlike `Number.parseInt` which
+ * would have silently truncated to `1` / `12`.
+ *
+ * Returns:
+ *   - `null` when the input is empty (= "no cap" / clear cap)
+ *   - a positive integer on valid input
+ *   - `undefined` on invalid input
+ */
+export function parseGoalTurnCapInput(value: string): number | null | undefined {
+  const trimmed = value.trim();
+  if (trimmed.length === 0) {
+    return null;
+  }
+  const parsed = Number(trimmed);
+  return Number.isSafeInteger(parsed) && parsed > 0 ? parsed : undefined;
+}
